+++ packages/orq-rc/src/models/operations/createchunk.ts
@@ -582,7 +582,7 @@ export const CreateChunkResponseBody$inboundSchema: z.ZodType<
   knowledge_id: z.string(),
   workspace_id: z.string(),
   data_source_id: z.string(),
-  id: z.string().default("chunk_01JQTF7J36ZBVBD9YTA9F6CDNJ"),
+  id: z.string().default("chunk_01JQTM2A46TPRZBYMFT3QHM9BB"),
   text: z.string(),
   enabled: z.boolean().default(true),
   metadata: z.lazy(() => CreateChunkKnowledgeMetadata$inboundSchema).optional(),
@@ -627,7 +627,7 @@ export const CreateChunkResponseBody$outboundSchema: z.ZodType<
   knowledgeId: z.string(),
   workspaceId: z.string(),
   dataSourceId: z.string(),
-  id: z.string().default("chunk_01JQTF7J36ZBVBD9YTA9F6CDNJ"),
+  id: z.string().default("chunk_01JQTM2A46TPRZBYMFT3QHM9BB"),
   text: z.string(),
   enabled: z.boolean().default(true),
   metadata: z.lazy(() => CreateChunkKnowledgeMetadata$outboundSchema)