+++ packages/orq-rc/RELEASES.md
@@ -1209,3 +1209,13 @@ Based on:
 - [typescript v3.3.0-rc.23] packages/orq-rc
 ### Releases
 - [NPM v3.3.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.23 - packages/orq-rc
+
+## 2025-04-02 06:48:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.527.1 (2.564.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.3.0-rc.24] packages/orq-rc
+### Releases
+- [NPM v3.3.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.24 - packages/orq-rc